+++ quantify/visualization/instrument_monitor.py
@@ -4,8 +4,11 @@
 # Copyright (C) Qblox BV & Orange Quantum Systems Holding BV (2020-2021)
 # -----------------------------------------------------------------------------
 import time
+import warnings
+
 import pyqtgraph as pg
 import pyqtgraph.multiprocess as pgmp
+from pyqtgraph.multiprocess.remoteproxy import NoResultError
 
 from qcodes.instrument.base import Instrument
 from qcodes.utils import validators as vals
@@ -14,8 +17,6 @@ from qcodes.instrument.parameter import ManualParameter
 from quantify.data.handling import snapshot
 from quantify.utilities.general import traverse_dict
 from quantify.visualization.ins_mon_widget import qc_snapshot_widget
-
-import warnings
 
 
 class InstrumentMonitor(Instrument):
@@ -40,9 +41,7 @@ class InstrumentMonitor(Instrument):
     proc = None
     rpg = None
 
-    def __init__(
-        self, name, window_size: tuple = (600, 600), remote: bool = True, **kwargs
-    ):
+    def __init__(self, name, window_size: tuple = (600, 600), remote: bool = True):
         """
         Initializes the pyqtgraph window
 
@@ -104,7 +103,17 @@ class InstrumentMonitor(Instrument):
         )  # pyqtgraph multiprocessing
         self.__class__.rpg = self.proc._import("pyqtgraph")
         qc_widget = "quantify.visualization.ins_mon_widget.qc_snapshot_widget"
-        self.__class__.rwidget = self.proc._import(qc_widget)
+
+        for i in range(5):
+            try:
+                self.__class__.rwidget = self.proc._import(qc_widget)
+            except NoResultError as exception:
+                # Try a few times before giving up
+                time.sleep(0.2)
+                if i == 4:
+                    raise exception
+            else:
+                break
 
     def create_widget(self, window_size: tuple = (1000, 600)):
         """