+++ tests/data/test_data_handling.py
@@ -2,14 +2,15 @@ import os
 import shutil
 import pytest
 import dateutil
+from datetime import datetime
+
 import xarray as xr
 import numpy as np
-from quantify.data.types import TUID
-import quantify.data.handling as dh
-from quantify.measurement.control import MeasurementControl
-from datetime import datetime
 from qcodes import ManualParameter
+from quantify.data.types import TUID
+from quantify.measurement.control import MeasurementControl
 from quantify.utilities._tests_helpers import get_test_data_dir
+import quantify.data.handling as dh
 
 
 test_datadir = get_test_data_dir()
@@ -35,7 +36,7 @@ def test_initialize_dataset():
     dataset = dh.initialize_dataset(setable_pars, setpoints, getable_pars)
 
     assert isinstance(dataset, xr.Dataset)
-    assert len(dataset.data_vars) == 2
+    assert len(dataset.data_vars) == 1
     assert dataset.attrs.keys() == {"tuid"}
     assert dataset.variables.keys() == {"x0", "y0"}
 
@@ -51,6 +52,9 @@ def test_initialize_dataset():
     assert y0.attrs["name"] == "y"
     assert y0.attrs["long_name"] == "Signal amplitude"
 
+    assert set(dataset.coords.keys()) == {"x0"}
+    assert set(dataset.dims.keys()) == {"dim_0"}
+
 
 def test_initialize_dataset_2D():
     xpar = ManualParameter("x", unit="m", label="X position")
@@ -64,9 +68,10 @@ def test_initialize_dataset_2D():
     dataset = dh.initialize_dataset(setable_pars, setpoints, getable_pars)
 
     assert isinstance(dataset, xr.Dataset)
-    assert len(dataset.data_vars) == 3
+    assert len(dataset.data_vars) == 1
     assert dataset.attrs.keys() == {"tuid"}
     assert set(dataset.variables.keys()) == {"x0", "x1", "y0"}
+    assert set(dataset.coords.keys()) == {"x0", "x1"}
 
 
 def test_getset_datadir():
@@ -369,3 +374,36 @@ def test_dynamic_dataset():
     assert not np.isnan(dset["x0"]).any()
     assert not np.isnan(dset["x1"]).any()
     assert not np.isnan(dset["y0"]).any()
+
+    assert "tuid" in set(dset.attrs)
+
+
+def test_to_gridded_dataset():
+    dh.set_datadir(test_datadir)
+    tuid = "20200504-191556-002-4209ee"
+    dset_orig = dh.load_dataset(tuid)
+    dset_gridded = dh.to_gridded_dataset(dset_orig)
+
+    assert dset_gridded.attrs["tuid"] == tuid
+    assert tuple(dset_gridded.dims.keys()) == ("x0", "x1")
+    assert tuple(dset_gridded.coords.keys()) == ("x0", "x1")
+    assert tuple(dset_gridded.dims.values()) == (50, 11)
+    assert dset_gridded["y0"].dims == ("x0", "x1")
+    assert len(dset_gridded["x0"].values) == len(np.unique(dset_orig["x0"]))
+
+    for var in ("x0", "x1", "y0"):
+        assert dset_orig[var].attrs == dset_gridded[var].attrs
+
+    y0 = dset_gridded["y0"].values
+
+    indices = [[10, 9], [22, 6], [7, 3], [10, 1], [18, 1], [0, 3]]
+    expected = [
+        -0.7983563142002691,
+        0.1436698700195456,
+        0.2493959207434933,
+        0.7983563142002691,
+        -0.6970549632987115,
+        -0.3999999999999999,
+    ]
+
+    assert [y0[tuple(idxs)] for idxs in indices] == expected